Delete false info on solid classes
Earlier I wrote that solids don't have class properties. This is no longer true, they currently extend the same class as point classes. They are functionality identical when it comes to properties. They are set the same way in the resource Class Properties dictionary, and set the same way in code with `if "keyname" in properties: print(properties.keyname)`